# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "一款面向小型团队的任务进度共享软件",
        "title": "Team",
        "termsOfService": "http://swagger.io/terrms",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:9918",
    "paths": {
...
        "/task/:task_id": {
            "delete": {
                "description": "\"删除一个任务\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "summary": "\"删除任务\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "task_id",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功"
                    },
                    "400": {
                        "description": "删除失败"
                    },
                    "401": {
                        "description": "身份验证失败"
                    }
                }
            }
        },
        "/task/pro_step_info/:team_id": {
            "get": {
                "description": "\"在新建任务界面选择项目后(填入项目名称)返回步骤\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "summary": "\"项目对应的步骤\"",
                "parameters": [
                    {
                        "description": "项目名称",
                        "name": "pro",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Project"
                        }
                    },
                    {
                        "type": "string",
                        "description": "team_id",
                        "name": "team_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "获取失败"
                    }
                }
            }
        },
        "/task/team_info/:team_id": {
            "get": {
                "description": "\"获取项目任务新建所需的团队成员和项目名称\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "summary": "\"获取任务编辑页面需要的信息\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "team_id",
                        "name": "team_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功"
                    },
                    "400": {
                        "description": "获取失败"
                    }
                }
            }
        },
...
        "/team/task/:task_id": {
            "get": {
                "description": "\"编辑前查看任务信息\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "project"
                ],
                "summary": "\"查看任务\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "task_id",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "获取失败"
                    },
                    "401": {
                        "description": "身份验证失败"
                    }
                }
            }
        },
        "/team/task/:task_id/:team_id": {
            "put": {
                "description": "\"创建人修改一个任务\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "project"
                ],
                "summary": "\"修改任务\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "task_id",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "填入需要修改的信息",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Task"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "修改成功"
                    },
                    "400": {
                        "description": "删除失败"
                    },
                    "401": {
                        "description": "身份验证失败"
                    }
                }
            }
        },
        "/team/task/:team_id": {
            "post": {
                "description": "\"项目任务新建\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "summary": "\"创建任务并分配\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "team_id",
                        "name": "step_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "注意填入的成员是结构体",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Task"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "创建失败"
                    },
                    "401": {
                        "description": "身份验证失败"
                    },
                    "404": {
                        "description": "格式错误"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Project": {
            "type": "object",
            "properties": {
                "create_time": {
                    "type": "string"
                },
                "creator_id": {
                    "type": "integer"
                },
                "deadline": {
                    "type": "string"
                },
                "project_id": {
                    "type": "integer"
                },
                "project_name": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "step": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "team_id": {
                    "type": "integer"
                }
            }
        },
        "model.Task": {
            "type": "object",
            "properties": {
                "createtime": {
                    "type": "string"
                },
                "creator_id": {
                    "type": "integer"
                },
                "deadline": {
                    "type": "string"
                },
                "member": {
                    "description": "以下不是必要的，但方便实用",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "project": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "step": {
                    "type": "string"
                },
                "step_id": {
                    "type": "integer"
                },
                "task_id": {
                    "type": "integer"
                },
                "team_id": {
                    "type": "integer"
                }
            }
        },
...
    }
}